use size_t for iterating to avoid sign-extenstions

+++ src/motor_message.cc
@@ -178,7 +178,7 @@ int MotorMessage::deserialize(const std::vector<uint8_t> &serialized){
 int MotorMessage::verifyType(uint8_t t) {
   //Return 1 good
   //Return 0 for bad
-  for (int i = 0; i < sizeof(valid_types) / sizeof(valid_types[0]); ++i)
+  for (size_t i = 0; i < sizeof(valid_types) / sizeof(valid_types[0]); ++i)
   {
     if (t == valid_types[i])
       return 1;
@@ -189,7 +189,7 @@ int MotorMessage::verifyType(uint8_t t) {
 int MotorMessage::verifyRegister(uint8_t r) {
   //Return 1 good
   //Return 0 for bad
-  for (int i = 0; i < sizeof(valid_registers) / sizeof(valid_registers[0]); ++i)
+  for (size_t i = 0; i < sizeof(valid_registers) / sizeof(valid_registers[0]); ++i)
   {
     if (r == valid_registers[i])
       return 1;
